Cannot delete car that have active or pending orders.

+++ Backend/Services/Services/CarManagementService.cs
@@ -28,11 +28,25 @@ namespace Services.Services {
             }
         }
 
-        public IActionResult DeleteCar(int ID) {
+        public IActionResult DeleteCar(int ID, bool hard = false) {
             using (var context = new DataContext(Config)) {
                 if (context.Cars.Any(x => x.ID == ID)) {
-                    var dbo = context.Cars.Single(x => x.ID == ID);
-                    context.Cars.Remove(dbo);
+                    var dbo = context.Cars.First(x => x.ID == ID);
+                    var orders = context.Orders.Where(x => x.Car.ID == dbo.ID);
+                    foreach (var order in orders) {
+                        if (OrderStateHelper.GetOrderState(order) == OrderStates.ACTIVE) {
+                            return new ObjectResult("User has active orders!") { StatusCode = StatusCodes.Status409Conflict };
+                        }
+                        if (OrderStateHelper.GetOrderState(order) == OrderStates.PENDING) {
+                            return new ObjectResult("User has pending orders!") { StatusCode = StatusCodes.Status409Conflict };
+                        }
+                    }
+                    if (hard) {
+                        context.Cars.Remove(dbo);
+                        context.Orders.RemoveRange(orders);
+                    } else {
+                        dbo.IsDeleted = true;
+                    }
                     context.SaveChanges();
                     return new StatusCodeResult(StatusCodes.Status200OK);
                 } else {
